Fixing a security issue
Github's automated security checker found an old version of a library that needed to be updated to avoid a security vulnerability. I may as well fix it, in case someone uses this bit of code as the basis for an actual application, even though it's really only a sample. (Better safe than sorry.)